+++ ta4j-core/src/main/java/org/ta4j/core/criteria/pnl/GrossReturnCriterion.java
@@ -53,5 +53,5 @@ public class GrossReturnCriterion extends AbstractReturnCriterion {
     protected Num calculateReturn(BarSeries series, Position position) {
         return position.getGrossReturn(series);
     }
-  
+
 }